+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Change Log
 
-All notable changes to the "polaris" extension will be documented in this file.
+## [0.7.0]
+- Add syntax highlighting for string escapes and interpolation
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## [0.5.0]
+- Add support for the Polaris Language Server (polarislsp)
 
 ## [0.1.0]
 - Map literals should not be highlighted as comments anymore